requirements.txt
aiofiles==0.5.0
aiosqlite==0.16.1
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.22.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.1
idna==2.10
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
orjson==3.5.1
promise==2.3
pydantic==1.8.1
pypika-tortoise==0.1.0
python-dotenv==0.16.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
tortoise-orm==0.17.1
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
